Require PolyChaos 2.1.2 for 32-bit numberPolynomials (#611)
* Temporarily source PolyChaos and QMC masters for 32-bit CI.


* Drop QMC git source now that 0.4.3 is registered.


* CI: use SciML/.github#137 so LTS can Pkg.add URL [sources].


* Require PolyChaos 2.1.2; drop temporary [sources] and #137 pin.

2.1.2 is on master (#193); registration pending. Cross-repo PRs
cannot use refs/pull/137/head as a reusable workflow ref.


* Fix PolyChaos UUID accidentally overwritten by compat bump.


* chore: retest after PolyChaos 2.1.2 registration.


* Use factorial(Int64(n)) in Lobachevsky kernels for 32-bit.

On i686, Int is Int32 and factorial(n) overflows the Int32 table for n≥13
(tests use n up to 20).


* CI: retrigger Tests after Lobachevsky Int64 factorial fix.


---------

+++ test/lobachevsky.jl
@@ -24,7 +24,7 @@ using Cubature
         @test_throws ArgumentError LobachevskySurrogate(x, y, 0.0, 2.0, n = 3)
         @test_throws ArgumentError LobachevskySurrogate(x, y, 0.0, 2.0, n = 0)
         @test_throws ArgumentError LobachevskySurrogate(x, y, 0.0, 2.0, n = -2)
-        # factorial(n) overflows Int64 past 20
+        # factorial(Int64(n)) overflows Int64 past 20
         @test_throws ArgumentError LobachevskySurrogate(x, y, 0.0, 2.0, n = 22)
         @test LobachevskySurrogate(x, y, 0.0, 2.0, n = 20) isa LobachevskySurrogate
         @test LobachevskySurrogate(x, y, 0.0, 2.0, alpha = 4.0) isa LobachevskySurrogate
@@ -77,7 +77,7 @@ using Cubature
                     val += (-1)^l * binomial(n, l) * a^(n - 1)
                 end
             end
-            return val * (c / (2^n * factorial(n - 1)))
+            return val * (c / (2^n * factorial(Int64(n) - 1)))
         end
         function direct_int(point, n)
             res = zero(eltype(point))
@@ -88,7 +88,7 @@ using Cubature
                     res = res + (-1)^k * binomial(n, k) * c^n
                 end
             end
-            return res / (2^n * factorial(n))
+            return res / (2^n * factorial(Int64(n)))
         end
 
         @test sqrt(12 / 3) == 2.0
